# server/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Food Delivery API",
    "description": "API for food delivery system including user, vendor, delivery agent, and analytics functionality",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5050/api/v1",
      "description": "Local development server"
    }
  ],
  "paths": {
...
"/auth/vendor/updateDetail": {
  "put": {
    "tags": ["Vendor"],
    "summary": "Update vendor details",
    "operationId": "updateVendorDetails",
    "security": [
      {
        "bearerAuth": []
      }
    ],
    "requestBody": {
      "content": {
        "application/json": {
          "schema": {
            "type": "object",
            "properties": {
              "firstname": {
                "type": "string",
                "example": "kevin"
              },
              "lastname": {
                "type": "string",
                "example": "patel"
              },
              "number": {
                "type": "integer",
                "example": 9876543210
              },
              "email": {
                "type": "string",
                "example": "kevin.doe@example.com"
              },
              "totalTables": {
                "type": "integer",
                "example": 20
              }
            }
          }
        }
      }
    },
    "responses": {
      "200": {
        "description": "Vendor details updated successfully",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": true
                },
                "message": {
                  "type": "string",
                  "example": "Vendor updated successfully!"
                },
                "token": {
                  "type": "string",
                  "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2ODEzMmU3NDViYTIyODQ5YzY0OTlkZWQiLCJlbWFpbCI6ImtldmluLmRvZUBleGFtcGxlLmNvbSIsInJvbGUiOiJ2ZW5kb3IiLCJpYXQiOjE3NDY2MTUxMzksImV4cCI6MTc0NjcwMTUzOX0.c7bDkMXM8vJbHF9Y8e2tOSIYT2cMWL-_qXxkL6i-m8M"
                },
                "vendor": {
                  "type": "object",
                  "properties": {
                    "userId": {
                      "type": "string",
                      "example": "68132e745ba22849c6499ded"
                    },
                    "firstname": {
                      "type": "string",
                      "example": "kevin"
                    },
                    "lastname": {
                      "type": "string",
                      "example": "patel"
                    },
                    "fullname": {
                      "type": "string",
                      "example": "kevin patel"
                    },
                    "number": {
                      "type": "integer",
                      "example": 9876543210
                    },
                    "email": {
                      "type": "string",
                      "example": "kevin.doe@example.com"
                    },
                    "totalTables": {
                      "type": "integer",
                      "example": 20
                    },
                    "isActive": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
},
...
"/auth/user/register": {
  "post": {
    "tags": ["Delivery Agent"],
    "summary": "Register a new user",
    "operationId": "registerUser",
    "requestBody": {
      "required": true,
      "content": {
        "application/json": {
          "schema": {
            "type": "object",
            "properties": {
              "firstname": {
                "type": "string",
                "example": "bhuvaaan"
              },
              "lastname": {
                "type": "string",
                "example": "bam"
              },
              "number": {
                "type": "integer",
                "example": 9876543210
              },
              "email": {
                "type": "string",
                "format": "email",
                "example": "bhuvaan.bam@example.com"
              }
            },
            "required": ["firstname", "lastname", "number", "email"]
          }
        }
      }
    },
    "security": [
      {
        "bearerAuth": []
      }
    ],
    "responses": {
      "200": {
        "description": "User registered successfully by vendor",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": true
                },
                "message": {
                  "type": "string",
                  "example": "User registered successfully by vendor"
                },
                "token": {
                  "type": "string",
                  "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                },
                "user": {
                  "type": "object",
                  "properties": {
                    "userId": {
                      "type": "string",
                      "example": "681b4c6a40bc3f6fb0d5be03"
                    },
                    "firstname": {
                      "type": "string",
                      "example": "bhuvaaan"
                    },
                    "lastname": {
                      "type": "string",
                      "example": "bam"
                    },
                    "email": {
                      "type": "string",
                      "example": "bhuvaan.bam@example.com"
                    },
                    "number": {
                      "type": "integer",
                      "example": 9876543210
                    }
                  }
                }
              }
            }
          }
        }
      },
      "400": {
        "description": "Bad Request - Missing or invalid fields"
      }
    }
  }
},
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}